requirements.txt
annotated-types==0.7.0
anyio==4.7.0
blinker==1.9.0
click==8.1.8
cloudpickle==3.1.0
contourpy==1.3.1
cycler==0.12.1
Farama-Notifications==0.0.4
fastapi==0.115.6
Flask==3.1.0
fonttools==4.55.3
gymnasium==1.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
numpy==2.2.1
packaging==24.2
pillow==11.1.0
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
scikit-learn==1.6.0
scipy==1.15.0
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
Werkzeug==3.1.3